requirements.txt
Django==2.2.10
djangorestframework==3.11.2
djangorestframework-simplejwt==4.6.0
PyJWT==2.0.0
pytz==2020.4
sqlparse==0.4.1
drf-yasg==1.17.1
